3th API update completed

+++ routes/cart/cart.controller.js
@@ -4,11 +4,10 @@ const error = require(path.join(__dirname, '../../helper/errorHandler'));
 const Cart = require(path.join(__dirname, './cart.model.js'));
 
 exports.getAllCartByMemId = (req, res, next) => {
-  console.log("get all cart with mem_id");
   const mem_id = req.params.mem_id;
 
   const respond = (results) => {
-    res.status(200).json(results);
+    results ? res.status(200).json(results) : next(error(400));
   };
 
   const onError = (err) => {
@@ -25,44 +24,50 @@ exports.getAllCartByMemId = (req, res, next) => {
 };
 
 exports.createCart = (req, res, next) => {
-  console.log("Create a cart");
 
-  const { mem_id, club_id } = req.body;
+  const createList = {
+    mem_id: req.body.mem_id,
+    club_id:  req.body.club_id
+  };
 
-  const respond = (result) => {    
-    res.status(200).json(result);
+  const respond = (result) => {
+    result ? res.status(200).json(result) : next(error(400));
   };
 
   const onError = (err) => {
     next(err);
   };
 
-  Cart.create({
-    mem_id: req.body.mem_id,
-    club_id: req.body.club_id
-  })
+  Cart.create(createList)
   .then(respond)
   .catch(onError);
 };
 
 exports.deleteCart = (req, res, next) => {
-  console.log("Remove a cart");
-
-  const cart_id = req.params.cart_id;
 
   const respond = (result) => {
-    res.status(201).json(result);
+    if(result) {
+      Cart.destroy({
+        where : {
+          cart_id : req.params.cart_id
+        }
+      })
+      .then(result => {
+        res.send(200);
+      })
+      .catch(err => {
+        next(err);
+      });
+    } else {
+      next(error(400));
+    }
   };
 
   const onError = (err) => {
     next(err);
   };
 
-  Cart.destroy({
-    where : {
-      cart_id : cart_id
-    }
-  })
+  Cart.findById(req.params.cart_id)
   .then(respond)
   .catch(onError);
 };
